Updated models and methods for creating a new transaction and paying it partially. It won't create a new transaction for every partial payment anymore. Created separate method to handle partial payment. - When the transaction of the total amount is made, an invoice is created in Document model. - If an amount is paid when the transaction is created, a receipt type document will be created as well in Documents model. - It will create a receipt in Documents model everytime an amount of transactionId x is paid.
Fixed the formula of remaining amount.

Created a RecurringTask class that inherits the BackgroundTask. Used it for checking all the transactions from database every day at 12a.m using timer and modifying the paymentstatus as overdue if the due date has passed and the transaction has not been fully paid.